resource_control: create the default resource group (#6206)
ref tikv/pd#4399, ref pingcap/tidb#42403

resource_control: create the default resource group

+++ pkg/mcs/resource_manager/server/manager.go
@@ -22,11 +22,11 @@ import (
 	"time"
 
 	"github.com/gogo/protobuf/proto"
-	"github.com/pingcap/errors"
 	"github.com/pingcap/failpoint"
 	rmpb "github.com/pingcap/kvproto/pkg/resource_manager"
 	"github.com/pingcap/log"
 	bs "github.com/tikv/pd/pkg/basicserver"
+	"github.com/tikv/pd/pkg/errs"
 	"github.com/tikv/pd/pkg/storage/endpoint"
 	"github.com/tikv/pd/pkg/storage/kv"
 	"github.com/tikv/pd/pkg/utils/logutil"
@@ -37,6 +37,9 @@ const (
 	defaultConsumptionChanSize = 1024
 	metricsCleanupInterval     = time.Minute
 	metricsCleanupTimeout      = 20 * time.Minute
+
+	reservedDefaultGroupName = "default"
+	middlePriority           = 8
 )
 
 // Manager is the manager of resource group.
@@ -121,6 +124,24 @@ func (m *Manager) Init(ctx context.Context) {
 		}
 	}
 	m.storage.LoadResourceGroupStates(tokenHandler)
+
+	// Add default group
+	defaultGroup := &ResourceGroup{
+		Name: reservedDefaultGroupName,
+		Mode: rmpb.GroupMode_RUMode,
+		RUSettings: &RequestUnitSettings{
+			RU: GroupTokenBucket{
+				Settings: &rmpb.TokenLimitSettings{
+					FillRate:   1000000,
+					BurstLimit: -1,
+				},
+			},
+		},
+		Priority: middlePriority,
+	}
+	if err := m.AddResourceGroup(defaultGroup); err != nil {
+		log.Warn("init default group failed", zap.Error(err))
+	}
 	// Start the background metrics flusher.
 	go m.backgroundMetricsFlush(ctx)
 	go func() {
@@ -136,7 +157,7 @@ func (m *Manager) AddResourceGroup(group *ResourceGroup) error {
 	_, ok := m.groups[group.Name]
 	m.RUnlock()
 	if ok {
-		return errors.New("this group already exists")
+		return errs.ErrResourceGroupAlreadyExists.FastGenByArgs(group.Name)
 	}
 	err := group.CheckAndInit()
 	if err != nil {
@@ -157,13 +178,13 @@ func (m *Manager) AddResourceGroup(group *ResourceGroup) error {
 // ModifyResourceGroup modifies an existing resource group.
 func (m *Manager) ModifyResourceGroup(group *rmpb.ResourceGroup) error {
 	if group == nil || group.Name == "" {
-		return errors.New("invalid group name")
+		return errs.ErrInvalidGroup
 	}
 	m.Lock()
 	curGroup, ok := m.groups[group.Name]
 	m.Unlock()
 	if !ok {
-		return errors.New("not exists the group")
+		return errs.ErrResourceGroupNotExists.FastGenByArgs(group.Name)
 	}
 
 	err := curGroup.PatchSettings(group)
@@ -175,6 +196,9 @@ func (m *Manager) ModifyResourceGroup(group *rmpb.ResourceGroup) error {
 
 // DeleteResourceGroup deletes a resource group.
 func (m *Manager) DeleteResourceGroup(name string) error {
+	if name == reservedDefaultGroupName {
+		return errs.ErrDeleteReservedGroup
+	}
 	if err := m.storage.DeleteResourceGroupSetting(name); err != nil {
 		return err
 	}
